fix: fixed problem with proof and details editors
In the finding creation request, there are no poc
and affected assets properties in the object. At
the finding creation handler object will not
contain these fields. Unfortunately database will
not add them to finding. So after switching tabs
we get undefined value for POC editor. It looks
safe to init it in empty string and editor would
be able to work with it.